Fixup Deep Sleep in combination with bluesleep.
Due to the inclusion of FMRadio, the wake mechanism in the userspace
moved from using the procfs to the ioctl assertion mechanism. This
commit makes the bluesleep and the msm uart driver fully aware of the
ioctl assertion mechanism. The basic idea is to trigger the bluesleep
functions that are called during assertion/deassertion via the procfs
from the msm_serial_hs driver whenever an assert/deassert ioctl is
delivered.

NOTE: This change requires modifications to the libbt configuration
to operate correctly (remove the wake via procfs).

In particular, the following changes have been made:

* Create a bluesleep header file to allow the msm_serial_hs driver to
access the bluesleep assert/deassert functions.

* msm_serial_hs: Call the assert/deassert functions
(bluesleep_outgoing_data() / bluesleep_tx_allow_sleep()) when the
corresponding ioctl commands are received.

* bcm43xx: Call the bluesleep_start/stop funtions when BT is enabled
or disabled.

* bluesleep: Make the bluesleep_tx_allow_sleep function externally
accessible.

* bluesleep: Introduce a state tracker variable for uart clock on/off
requests and use it in the hsuart_power method. Further, remove any
direct invocations of msm_hs_request_clock_on/off.

* bluesleep: Add a state flag to the bluesleep_start function to signal
the context from which the bluesleep driver is called: 0 indicates that
no clock has been asserted before, while 1 signals the opposite. This
flag is required to correctly set the initial state which enables deep
sleep and FMRadio with the current limitations. This flag should be no
longer needer once the FMRadio driver uses the standard ldisc.

* bluesleep: bluesleep_tx_allow_sleep: for the ldisc case, set the
tx_timer instead of issueing the sleep worker directly to avoid sending
BT to sleep during pairing.

Change-Id: I50eb35b6af93f6bebb2122b3f80061f04b120998
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>